Fix for loadPresets not updating the Flags List
loadPreset() updates the dictionaries, but it does not clear and update the self.flags() list.

As a result, the FlagString is not updated in updateFlagString() because updateFlagString() only looks at the self.flags list.

More importantly, the generateSeed() method passes the contents of self.flags to the Randomizer.randomize() method, meaning that if a user was to load a preset and then attempt to randomize the ROM, it won't be using the correct flags (because self.flags was never updated).

This fix iterates over the dictionaries after the dictionaries have been updated. Why not update self.flags during the same loop used to update the dictionaries? Because a set is used to update the dictionaries. Updating self.flags afterwards ensures that self.flags is sorted both alphabetically and by dictionary.