Fix for loadPreset not updating flags list
When a preset is selected, loadPreset() updates the dictionaries, but it does not clear and update the self.flags() list. This causes a couple of problems:

1) The flag string on the UI is not properly updated in updateFlagString() because self.flagString is updated using the self.flags list.
2) More importantly, the generateSeed() method passes the contents of self.flags to the Randomizer.randomize() method. If a user loads a preset and then attempts to randomize the ROM, it won't be using the correct flags (because self.flags was never updated).

This fix iterates over the dictionaries after the dictionaries have been updated. It was possible to update self.flags during the same loops that update the dictionaries, but doing so would not have produced a flags list as neatly sorted because the dictionaries are updated using a set.