fix bug in appearance selection (?)
this fix has been rotting in a branch since 2019. I looked into it and it seems like, if you had Terra swapped to vanilla Strago, for example, and not part of makeover, makeover would cull *Terra* clone sprites instead of *Strago* clone sprites. This .. should fix that issue? I'm pretty sure? Has that been an unreported bug for this long?