GP Code and QSpinBox bug fixes.
Fixed GP code not working. The values were not being written to the ROM.

Sometimes QSpinBox values would be like 1.899999 instead of 1.9. Because the program denotes "Off" by a value of -1.0, sometimes the value would end up as -0.99999, which failed to turn off the code. Rounding fixes this and ensures the values used to randomize match what is seen in the GUI.