Reworked file handling, threading
Previously, the randomizer would open a source ROM, make a copy of it, then open and close both the source and destination ROM files multiple times throughout the randomization process.

With this change, the source ROM is opened and two in-memory bytestreams are created from class io.BytesIO. These streams are passed around the randomizer and written to in the same way as the previous file streams, except the BytesIO streams are never closed at any point except the very end of randomization, where the destination ROM is finally written to disk.

There may be performance advantages to this change (randomization might happen more quickly), but the real reason is to make it easier for The Secret Project to function. The change should be invisible from the user's perspective.

Other changes:
- Paths that were previously relative are now absolute, calculated using pathlib.
- Replaced customthreadpool implementation with multiprocessing Process and Pipe objects. Pipes are used to allow communication between the GUI and randomize() processes and between randomize() and remonsterate processes.
- Swapped the positions of the Cancel and Confirm buttons on the GUI seed generation confirmation. My muscle memory told me they were backwards.
- Added output directory write permission check before the randomization process begins.
- Slightly adjusted GUI exception message.