Fix capitalization in "scenarionottaken" Shadow dialogue.
"Shadow: Besting an Imperial Soldier..."
->
"Shadow: Besting an Imperial soldier..."

"Locke: So the empire hired you to take me out, is that it?"
->
"Locke: So the Empire hired you to take me out, is that it?"

"Shadow: This town reeks of Imperial Occupation. I find it makes my drink taste sour."
->
"Shadow: This town reeks of Imperial occupation. I find it makes my drink taste sour."

"Shadow: The reaper is always just a step behind me..."
->
"Shadow: The Reaper is always just a step behind me..."

Capitalizing Reaper is not just grammatically correct (refers to Grim Reaper, a proper noun), but this also matches the original game's Shadow dialogue.  The vanilla game has this same line with capital R.