Bug Fix: Spinbox Flags
Fixed an issue where QDoubleSpinBox flags were not disabling properly because, when detecting if they should be disabled, an invalid comparison was used. 1 != 1.00.

Fixed an issue where QSpinBox and QDoubleSpinBox would become disabled when the value is set to 0, because 0 is a falsy value. Instead of using "if flag.value", we needed to use "if flag.value == ''".