Fixed item break for Life fizzling on dead chars.
Breaking an item that casts Life, Life 2, Phoenix or ChokeSmoke on a
dead character caused the spell to fizzle.  This commit fixes that by
checking the spell's flags to see whether it is permitted to target dead
characters, and allowing such targeting if so.

Because it's checking the spell data table, if BC in the future messes
with spells, this code will adjust automatically to what BC changes the
table to say.