Fixed the use of built-in presets in console mode.
The built-in presets having uppercase letters caused it to break when
entering those names, because lower() was being called prior to
comparing the names.

A comment in config.py said that preset names should always be
lowercase, probably because of this problem.  Rather than lowercase the
names, I fixed the code to work fine with uppercase preset names, and
removed the comment of course.

I fixed another bug in console too: presets with a space at the end of
their flag list would cause an error because split(' ') was used and it
tried to interpret an empty flag name.  I made the code filter empty
flag names, which happens if there are spaces at the beginning or end of
a flag list, or there is double spacing between flags.  Now they'll just
work as expected, ignoring the superfluous spacing.